LayoutTests/webaudio/prefixed-pannernode-basic-expected.txt

Basic tests for WebKit Prefixed PannerNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PannerNode has one input.
PASS PannerNode has one output.
PASS refDistance default value is 1.
PASS refDistance value is set to 270.52.
PASS maxDistance default value is 10000.
PASS maxDistance value is set to 100.55.
PASS rolloffFactor default value is 1.
PASS rolloffFactor value is set to 0.83.
PASS coneInnerAngle default value is 360.
PASS coneInnerAngle value is set to 240.45.
PASS coneOuterAngle default value is 360.
PASS coneOuterAngle value is set to 166.66.
PASS coneOuterGain default value is 0.
PASS coneOuterGain value is set to 0.35.
PASS PannerNode defaults to 'HRTF' panningModel.
PASS PannerNode defaults to 'inverse' distanceModel.
PASS panningModel: 'equalpower' is settable.
PASS panningModel: 'HRTF' is settable.
PASS distanceModel: 'linear' is settable.
PASS distanceModel: 'inverse' is settable.
PASS distanceModel: 'exponential' is settable.
PASS Setting .distanceModel to illegal string value did not throw an exception.
PASS Setting .distanceModel to illegal type did not throw an exception.
PASS successfullyParsed is true

TEST COMPLETE
